Basic working implementation for generating kotlin bindings
This is a basic implementation of a plugin for generating kotlin bindings. When applied, the plugin will create a KotlinGenTask that you can run on the source files.

The gen task itself visits the parsed Java files' AST to retrieve all the pertinent information (imports, methods, package, etc) and then generates kotlin bindings for those methods.

This is a working solution for the current files, but probably fragile and prone to breaking on future ones. Now time to improve it. #someonePLEASEmakeKotlinPoet