Simplify KotlinGenTask configuration
SourceTask can do the heavy lifting for us as far as include/exclude patterns, we just need to specify the source directories and those patterns on the set. The filters are applied when we call getSource() in the task itself. This saves a little time in configuration as well, which is always a plus