Add listener to target view before registering unsubscribe action.
This avoids a theoretically possible (albeit technically implausible) leak which could happen if the new subscribe was unsubscribed on a different thread between what was previously the unsubscription action registration and then the listener adding. If this happened, the unsubscribe action would run immediately and then the listener would be added to the target view with no way to remove it. Events would not be emitted because of the isUnsubscribed() checks in the callbacks.