Introduce notion of associatedImports for type resolution
JavaParser doesn't yield us the fully qualified class name for types. Historically we got away with this because we copied in the imports from the target class directly, but with KotlinPoet that's not an option anymore. This works around it by capturing the imports upfront and making a simple mapping of their simple name to ClassName and leverages them if they're present for kotlin type resolution.