Always emit editor events
Sometimes it is desireable to use an editor event without considering it
to be handled. For example when you want an EditText's send action to
perform some action and still close the keyboard.

The predicate passed to the RxTextView::editorActions and
RxTextView::editorActionEvents can be used to determine whether the
event should be considered to have been handled. A simple observable
filter can be used to only consider certain events.